+++ .rubocop.yml
@@ -1,5 +1,3 @@
-inherit_from: .rubocop_todo.yml
-
 require:
   - rubocop-rspec
   - rubocop-performance
@@ -7,8 +5,8 @@ require:
 AllCops:
   NewCops: enable
 
-# Layout/LineLength:
-#   Max: 120
+Layout/LineLength:
+  Max: 120
 
 Style/Documentation:
   Enabled: false
