--- .rubocop_todo.yml
@@ -1,53 +1,0 @@
-# This configuration was generated by
-# `rubocop --auto-gen-config`
-# on 2021-05-06 17:01:32 UTC using RuboCop version 1.14.0.
-# The point is for the user to remove these configuration records
-# one by one as the offenses are removed from the code base.
-# Note that changes in the inspected code, or installation of new
-# versions of RuboCop, may require this file to be generated again.
-
-# Offense count: 5
-# Cop supports --auto-correct.
-# Configuration parameters: AutoCorrect, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
-# URISchemes: http, https
-Layout/LineLength:
-  Max: 249
-
-# Offense count: 6
-# Configuration parameters: IgnoredMethods, CountRepeatedAttributes.
-Metrics/AbcSize:
-  Max: 34
-
-# Offense count: 1
-# Configuration parameters: IgnoredMethods.
-Metrics/CyclomaticComplexity:
-  Max: 9
-
-# Offense count: 7
-# Configuration parameters: CountComments, CountAsOne, ExcludedMethods, IgnoredMethods.
-Metrics/MethodLength:
-  Max: 38
-
-# Offense count: 1
-# Configuration parameters: IgnoredMethods.
-Metrics/PerceivedComplexity:
-  Max: 9
-
-# Offense count: 2
-# Cop supports --auto-correct.
-Performance/Casecmp:
-  Exclude:
-    - 'telegram_webhooks/searcher.rb'
-
-# Offense count: 2
-# Configuration parameters: AllowedChars.
-# AllowedChars: ©
-Style/AsciiComments:
-  Exclude:
-    - 'telegram_webhooks/app.rb'
-
-# Offense count: 1
-# Cop supports --auto-correct.
-Style/IfUnlessModifier:
-  Exclude:
-    - 'telegram_webhooks/app.rb'
