--- save_logs/Gemfile
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-source 'https://rubygems.org'
-
-ruby '~> 2.7.0'
