+++ telegram_webhooks/data_source/base.rb
@@ -18,6 +18,10 @@ module DataSource
       def s3
         @s3 ||= Aws::S3::Client.new(region: 'eu-north-1')
       end
+
+      def dynamodb
+        @dynamodb ||= Aws::DynamoDB::Client.new(region: 'eu-north-1')
+      end
     end
   end
 end
